e_commerce/requirements.txt
# requirements.txt
pymongo==4.5
# pydantic==2.3 #No funciona con Ninja
requests==2.31
email-validator==2.0
typing-extensions==4.8
django==4.2
django-bootstrap5
django-cors-headers==4.3.1
django-ninja-extra==0.19.8		
